chore(SimulationSettings): made delta_time a float instead of double

+++ src/SimulationScreen.hpp
@@ -27,7 +27,7 @@ struct GridSettings {
 
 // tunable values that come from adjust.h and may be re-read every frame
 struct SimulationSettings {
-    int& delta_time;
+    float& delta_time;
     int& substeps_per_frame;
     float& selected_sensitivity;
     float& objects_scale;
